+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for `sui_getBalance` [enhancement](https://github.com/FrankC01/pysui/issues/46). This returns
+total balance per coin type (i.e. `0x2::sui::SUI`) and is more efficient than getting all coins and
+totalling.
+
 ### Fixed
 
 ### Changed